+++ src/ui/main/mod.rs
@@ -8,6 +8,7 @@ use gtk::glib::clone;
 mod repair_game;
 mod download_wine;
 mod install_dxvk;
+mod install_dx12;
 mod create_prefix;
 mod download_diff;
 mod migrate_folder;
@@ -416,7 +417,8 @@ impl SimpleComponent for App {
                                                     Some(LauncherState::FolderMigrationRequired { .. }) |
                                                     Some(LauncherState::WineNotInstalled) |
                                                     Some(LauncherState::PrefixNotExists) |
-                                                    Some(LauncherState::DxvkNotInstalled) => "document-save-symbolic",
+                                                    Some(LauncherState::DxvkNotInstalled) |
+                                                    Some(LauncherState::Dx12NotInstalled) => "document-save-symbolic",
 
                                                     Some(LauncherState::GameUpdateAvailable(_)) |
                                                     Some(LauncherState::GameNotInstalled(_)) |
@@ -441,6 +443,7 @@ impl SimpleComponent for App {
                                                     Some(LauncherState::WineNotInstalled) => tr!("download-wine"),
                                                     Some(LauncherState::PrefixNotExists)  => tr!("create-prefix"),
                                                     Some(LauncherState::DxvkNotInstalled) => tr!("install-dxvk"),
+                                                    Some(LauncherState::Dx12NotInstalled) => tr!("install-dx12"),
 
                                                     Some(LauncherState::GameUpdateAvailable(diff)) |
                                                     Some(LauncherState::GameOutdated(diff)) |
@@ -1324,6 +1327,10 @@ impl SimpleComponent for App {
 
                     LauncherState::DxvkNotInstalled => {
                         install_dxvk::install_dxvk(sender, self.progress_bar.sender().to_owned())
+                    }
+
+                    LauncherState::Dx12NotInstalled => {
+                        install_dx12::install_dx12(sender, self.progress_bar.sender().to_owned())
                     }
 
                     LauncherState::GameUpdateAvailable(diff)